+++ Laptop/insurance Contract.h
@@ -18,7 +18,7 @@ void DailyUpdateOfInsuredMercs();
 //void InsuranceContractPayLifeInsuranceForDeadMerc( LIFE_INSURANCE_PAYOUT *pPayoutStruct );
 
 BOOLEAN AddLifeInsurancePayout( SOLDIERTYPE *pSoldier );
-void InsuranceContractPayLifeInsuranceForDeadMerc( UINT8	ubPayoutID );
+void InsuranceContractPayLifeInsuranceForDeadMerc( UINT16	ubPayoutID );
 void StartInsuranceInvestigation( UINT8	ubPayoutID );
 void EndInsuranceInvestigation( UINT8	ubPayoutID );
