+++ Strategic/Merc Contract.h
@@ -68,8 +68,8 @@ void FindOutIfAnyMercAboutToLeaveIsGonnaRenew( void );
 void BeginContractRenewalSequence( );
 void HandleContractRenewalSequence( );
 void EndCurrentContractRenewal( );
-void HandleMercIsWillingToRenew( UINT8 ubID );
-void HandleMercIsNotWillingToRenew( UINT8 ubID );
+void HandleMercIsWillingToRenew( UINT16 ubID );
+void HandleMercIsNotWillingToRenew( UINT16 ubID );
 
 BOOLEAN ContractIsExpiring( SOLDIERTYPE *pSoldier );
 UINT32 GetHourWhenContractDone( SOLDIERTYPE *pSoldier );
