+++ Strategic/Strategic Merc Handler.h
@@ -6,8 +6,8 @@ class SOLDIERTYPE;
 
 void StrategicHandlePlayerTeamMercDeath( SOLDIERTYPE *pSoldier );
 void MercDailyUpdate();
-void MercsContractIsFinished( UINT8	ubID );
-void RPCWhineAboutNoPay( UINT8	ubID );
+void MercsContractIsFinished( UINT16 ubID );
+void RPCWhineAboutNoPay( UINT16 ubID );
 void MercComplainAboutEquipment( UINT8 ubProfileID );
 BOOLEAN SoldierHasWorseEquipmentThanUsedTo( SOLDIERTYPE *pSoldier );
 void UpdateBuddyAndHatedCounters( void );
