+++ Strategic/Strategic Mines.cpp
@@ -136,7 +136,7 @@ UINT8 gubMonsterMineInfestation[]={
 // the static NPC dialogue faces
 //extern UINT32 uiExternalStaticNPCFaces[];
 extern FACETYPE	*gpCurrentTalkingFace;
-extern UINT8			gubCurrentTalkingID;
+extern UINT16			gubCurrentTalkingID;
