+++ Strategic/Strategic Movement.h
@@ -52,7 +52,7 @@ typedef struct WAYPOINT
 typedef struct PLAYERGROUP
 {
 	UINT8 ubProfileID;						//SAVE THIS VALUE ONLY.	The others are temp (for quick access)
-	UINT8 ubID;										//index in the Menptr array
+	UINT16 ubID;										//index in the Menptr array
 	SOLDIERTYPE *pSoldier;				//direct access to the soldier pointer
 	UINT8 bFlags;									//flags referring to individual player soldiers
 	struct PLAYERGROUP *next;			//next player in list
